Changed file locations.
I have added a file location for the Artichoke Dip (I wish that I had
chosen a shorter name).
I have also removed the redundant README after learning that terminal
is apparently case sensitive. I guess that it's better to learn that now
than it is to learn that later.